Django Rest Framework Projects/Demo Rest API/demo/views.py
from rest_framework.decorators import api_view
from rest_framework.response import Response
from .models import Post
from .serializers import PostSerializer


@api_view(["GET"])
def Overview(request):
    return Response({
        "This is an overview"
    })


@api_view(["GET"])
def GetPosts(request):
    post = Post.objects.all()
    serializer = PostSerializer(post, many=True)
    return Response(serializer.data)


@api_view(["GET"])
def GetPost(request, pk):
    post = Post.objects.get(id=pk)
    serializer = PostSerializer(post, many=False)
    return Response(serializer.data)


@api_view(["POST"])
def PostPost(request):
    serializer = PostSerializer(data=request.data)
    if serializer.is_valid():
        serializer.save()
    return Response(serializer.data)


@api_view(["GET", "PUT"])
def PutPost(request, pk):
    post = Post.objects.get(id=pk)
    serializer = PostSerializer(instance=post, data=request.data)
    if serializer.is_valid():
        serializer.save()
    return Response(serializer.data)


@api_view(["GET", "DELETE"])
def DeletePost(request, pk):
    post = Post.objects.get(id=pk)
    post.delete()
    return Response("Post Deleted Successfully")