Django Rest Framework Projects/Demo Rest API/requirements.txt
asgiref==3.3.1
Django==3.1.4
django-cors-headers==3.5.0
djangorestframework==3.12.2
gunicorn==20.0.4
pytz==2020.4
sqlparse==0.4.1
whitenoise==5.2.0
django-heroku==0.3.1
